Restrict student operations to teachers role

+++ src/student/student.controller.ts
@@ -1,9 +1,12 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, Put } from '@nestjs/common';
+import { Controller, Get, Post, Body, Param, Delete, UseGuards, Put, UnauthorizedException } from '@nestjs/common';
 import { StudentService } from './student.service';
 import { CreateStudentDto } from './dto/create-student.dto';
 import { UpdateStudentDto } from './dto/update-student.dto';
 import { AuthGuard } from '@nestjs/passport';
 import { ApiBearerAuth, ApiCreatedResponse, ApiOkResponse, ApiTags, ApiUnauthorizedResponse } from '@nestjs/swagger';
+import { GetUser } from '../auth/decorators/get-user-decorator';
+import { User } from '../auth/entities/user.entity';
+import { Role } from '../auth/enums/user-role.enum';
 
 @Controller('students')
 @UseGuards(AuthGuard('jwt'))
@@ -15,35 +18,55 @@ export class StudentController {
   @Post()
   @ApiCreatedResponse()
   @ApiUnauthorizedResponse()
-  create(@Body() createStudentDto: CreateStudentDto) {
+  create(@Body() createStudentDto: CreateStudentDto, @GetUser() user: User) {
+    if (user.role !== Role.TEACHER) {
+      throw new UnauthorizedException();
+    }
+
     return this.studentService.create(createStudentDto);
   }
 
   @Get()
   @ApiOkResponse()
   @ApiUnauthorizedResponse()
-  findAll() {
+  findAll(@GetUser() user: User) {
+    if (user.role !== Role.TEACHER) {
+      throw new UnauthorizedException();
+    }
+
     return this.studentService.findAll();
   }
 
   @Get(':id')
   @ApiOkResponse()
   @ApiUnauthorizedResponse()
-  findOne(@Param('id') id: string) {
+  findOne(@Param('id') id: string, @GetUser() user: User) {
+    if (user.role !== Role.TEACHER) {
+      throw new UnauthorizedException();
+    }
+
     return this.studentService.findOne(+id);
   }
 
   @Put(':id')
   @ApiOkResponse()
   @ApiUnauthorizedResponse()
-  update(@Param('id') id: string, @Body() updateStudentDto: UpdateStudentDto) {
+  update(@Param('id') id: string, @Body() updateStudentDto: UpdateStudentDto, @GetUser() user: User) {
+    if (user.role !== Role.TEACHER) {
+      throw new UnauthorizedException();
+    }
+
     return this.studentService.update(+id, updateStudentDto);
   }
 
   @Delete(':id')
   @ApiOkResponse()
   @ApiUnauthorizedResponse()
-  remove(@Param('id') id: string) {
+  remove(@Param('id') id: string, @GetUser() user: User) {
+    if (user.role !== Role.TEACHER) {
+      throw new UnauthorizedException();
+    }
+
     return this.studentService.remove(+id);
   }
 }
